OpenID4VP: Move internal classes to correct package
