+++ vck-rqes/src/commonTest/kotlin/at/asitplus/wallet/lib/rqes/SignatureRequestParsingTests.kt
@@ -2,7 +2,7 @@ package at.asitplus.wallet.lib.rqes
 
 import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.rqes.SignatureRequestParameters
-import at.asitplus.wallet.lib.oidc.helper.RequestParser
+import at.asitplus.wallet.lib.openid.RequestParser
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.types.shouldBeInstanceOf
